Move effects into sub-module
